fix: 배팅 금액과 이익을 long 자료형으로 변경
- 블랙잭 시 1원이면 1원이 되는 경우가 있어서 최소 배팅금액 100원으로 인상

+++ src/main/java/controller/BettingPhase.java
@@ -11,7 +11,7 @@ public class BettingPhase implements GamePhase {
     @Override
     public void execute(GameContext gameContext) {
         for (Player player : gameContext.getPlayers()) {
-            int bettingMoney = InputView.readPlayerBettingMoney(player.getName());
+            long bettingMoney = InputView.readPlayerBettingMoney(player.getName());
             player.setMoney(bettingMoney);
         }
     }
